+++ react/RatingSummary.tsx
@@ -1,70 +1,77 @@
-import React, {
-  FunctionComponent,
-  useContext,
-  useState,
-  useEffect,
-} from 'react'
+import React, { FunctionComponent, useContext } from 'react'
+import { FormattedMessage } from 'react-intl'
 import { ProductContext } from 'vtex.product-context'
 import queryRatingSummary from './graphql/queries/queryRatingSummary.gql'
-import getConfig from './graphql/getConfig.graphql'
-import { withApollo, graphql, ChildProps } from 'react-apollo'
+import getConfig from './graphql/getConfig.gql'
+import { withApollo, graphql, ChildProps, Query } from 'react-apollo'
 import { Link } from 'vtex.render-runtime'
 
 interface Settings {
-  appKey: string,
-  uniqueId: string,
-  merchantId: string,
+  appKey: string
+  uniqueId: string
+  merchantId: string
   merchantGroupId: string
 }
-const withSettings = graphql<{client: any}, Settings>(getConfig, { options: () => ({ ssr: false }) })
+const withSettings = graphql<{ client: any }, Settings>(getConfig, {
+  options: () => ({ ssr: false }),
+})
 
-const Reviews: FunctionComponent<ChildProps<Partial<ReviewProps>, Settings>> = props => {
+const RatingSummary: FunctionComponent<
+  ChildProps<Partial<RatingSummaryProps>, Settings>
+> = props => {
   const { product } = useContext(ProductContext)
 
-  const [average, setAverage] = useState(0)
-  const [totalReviews, setTotalReviews] = useState(0)
-  const [alreadyReviews, setAlreadyReviews] = useState(false)
+  const writeReviewLink =
+    props.data && props.data.getConfig
+      ? `/new-review?pr_page_id=${
+          product[props.data.getConfig.uniqueId]
+        }&pr_merchant_id=${props.data.getConfig.merchantId}&pr_api_key=${
+          props.data.getConfig.appKey
+        }&pr_merchant_group_id=${props.data.getConfig.merchantGroupId}`
+      : ''
 
-  useEffect(() => {
-    if (!product) {
-      return
-    }
+  return (
+    <Query
+      query={queryRatingSummary}
+      variables={{
+        sort: 'Newest',
+        page: 0,
+        pageId: JSON.stringify({
+          linkText: product.linkText,
+          productId: product.productId,
+          productReference: product.productReference,
+        }),
+        filter: 0,
+      }}
+    >
+      {({ data, loading }: { data: any; loading: boolean }) => {
+        const rollup =
+          data && data.productReviews && data.productReviews.results[0].rollup
+        const rating = rollup ? rollup.average_rating : 0
+        const numberOfReviews = rollup ? rollup.review_count : 0
 
-    const getReviews = (orderBy: any, page: any) => {
-      props.client
-        .query({
-          query: queryRatingSummary,
-          variables: {
-            sort: orderBy,
-            page: page || 0,
-            pageId: JSON.stringify({
-              linkText: product.linkText,
-              productId: product.productId,
-              productReference: product.productReference,
-            }),
-            filter: 0,
-          },
-        })
-        .then((response: any) => {
-          console.log('RESPONSE: ', response)
-          let reviews = response.data.productReviews.results[0].reviews // revisar se sempre vem 1 item nesse array
-          let rollup = response.data.productReviews.results[0].rollup
+        return (
+          <Summary
+            loading={loading}
+            writeReviewLink={writeReviewLink}
+            rating={rating}
+            numberOfReviews={numberOfReviews}
+          />
+        )
+      }}
+    </Query>
+  )
+}
 
-          setAverage(rollup != null ? rollup.average_rating : 0)
-          setAlreadyReviews(reviews.length ? true : false)
-          setTotalReviews(reviews.length ? reviews.length : 0)
-        })
-        .catch((error: any) => {
-          console.log('ERROR: ', error)
-        })
-    }
-
-    getReviews('Newest', 0)
-  }, [product, props.client])
-
-  return alreadyReviews ? (
+const Summary: FunctionComponent<SummaryProps> = ({
+  writeReviewLink,
+  loading,
+  rating,
+  numberOfReviews,
+}) => {
+  return (
     <div className="review__rating mw8 center mb5">
-      <div className="review__rating--stars dib relative v-mid mr2">
+      <div className="review__rating--stars dib relative v-mid mr4">
         <div className="review__rating--inactive nowrap">
           {[0, 1, 2, 3, 4].map((_, i) => {
             return i <= 3 ? (
@@ -101,7 +108,7 @@ const Reviews: FunctionComponent<ChildProps<Partial<ReviewProps>, Settings>> = p
         </div>
         <div
           className="review__rating--active nowrap overflow-hidden absolute top-0-s left-0-s"
-          style={{ width: average * 20 + '%' }}
+          style={{ width: rating * 20 + '%' }}
         >
           {[0, 1, 2, 3, 4].map((_, i) => {
             // let { average } = state;
@@ -113,7 +120,7 @@ const Reviews: FunctionComponent<ChildProps<Partial<ReviewProps>, Settings>> = p
                 xmlns="http://www.w3.org/2000/svg"
                 width="20"
                 height="20"
-                fill={average > i ? '#fc0' : '#eee'}
+                fill={rating > i ? '#fc0' : '#eee'}
                 viewBox="0 0 14.737 14"
               >
                 <path
@@ -127,7 +134,7 @@ const Reviews: FunctionComponent<ChildProps<Partial<ReviewProps>, Settings>> = p
                 xmlns="http://www.w3.org/2000/svg"
                 width="20"
                 height="20"
-                fill={average > i ? '#fc0' : '#eee'}
+                fill={rating > i ? '#fc0' : '#eee'}
                 viewBox="0 0 14.737 14"
               >
                 <path
@@ -139,77 +146,33 @@ const Reviews: FunctionComponent<ChildProps<Partial<ReviewProps>, Settings>> = p
           })}
         </div>
       </div>
-      <span className="review__rating--average dib v-mid c-muted-2 f6">
-        ({totalReviews})
+      <span className="review__rating--average mr4 dib c-muted-2 t-body f6-s">
+        (
+        <FormattedMessage
+          id="store/power-reviews.numberOfReviews"
+          values={{ number: numberOfReviews }}
+        />
+        )
       </span>
-      {!props.data.loading ? (
-        <Link
-          className='ml5 c-on-base dib f6'
-          to={`/new-review?pr_page_id=${
-            product[
-              props.data.getConfig.uniqueId
-            ]
-          }&pr_merchant_id=${
-            props.data.getConfig.merchantId
-          }&pr_api_key=${
-            props.data.getConfig.appKey
-          }&pr_merchant_group_id=${
-            props.data.getConfig.merchantGroupId
-          }`}
-        >
-          Write a review
-        </Link>
-      ) : null}
-    </div>
-  ) : (
-    <div className="review__rating mw8 center mb5">
-      <div className="review__rating--stars dib relative v-mid mr2">
-        <div className="review__rating--inactive nowrap">
-          {[0, 1, 2, 3, 4].map((_, i) => {
-            return (
-              <svg
-                key={i}
-                xmlns="http://www.w3.org/2000/svg"
-                width="20"
-                height="20"
-                fill="#eee"
-                viewBox="0 0 14.737 14"
-              >
-                <path
-                  d="M7.369,11.251,11.923,14,10.714,8.82l4.023-3.485-5.3-.449L7.369,0,5.3,4.885,0,5.335,4.023,8.82,2.815,14Z"
-                  transform="translate(0)"
-                ></path>
-              </svg>
-            )
-          })}
-        </div>
-      </div>
-      <span className="review__rating--average dib v-mid c-muted-2 f6">(0)</span>
-      {!props.data.loading ? (
-        <Link
-          className='ml5 c-on-base dib f6'
-          to={`/new-review?pr_page_id=${
-            product[
-              props.data.getConfig.uniqueId
-            ]
-          }&pr_merchant_id=${
-            props.data.getConfig.merchantId
-          }&pr_api_key=${
-            props.data.getConfig.appKey
-          }&pr_merchant_group_id=${
-            props.data.getConfig.merchantGroupId
-          }`}
-        >
-          Write a review
+      {!loading ? (
+        <Link className="dib c-on-base t-body f6-s" to={writeReviewLink}>
+          <FormattedMessage id="store/power-reviews.writeAReview" />
         </Link>
       ) : null}
     </div>
   )
 }
 
-interface ReviewProps {
-  client: any,
+interface SummaryProps {
+  writeReviewLink: string
+  rating: number
+  numberOfReviews: number
+  loading: boolean
+}
+
+interface RatingSummaryProps {
+  client: any
   data: any
 }
 
-export default withApollo(withSettings(Reviews))
+export default withApollo(withSettings(RatingSummary))