+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Use product-summary-context instead of deprecated product-summary.
+- Number of reviews in RatingSummary.
 
 ## [1.0.6] - 2019-07-02